backend/pyproject.toml
[project]
name = "llm-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.114.2",
    "langchain-openai>=0.2.11",
    "langchain>=0.3.9",
    "pydantic-settings",
    "sqlmodel>=0.0.22",
    "psycopg[binary]>=3.2.3",
    "alembic>=1.14.0",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "loguru>=0.7.3",
    "uvicorn==0.34.0",
]
